servidor/requirements.txt
Flask==2.0.0
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Cors==4.0.2
Flask-Admin==1.6.0
PyJWT==2.6.0
Werkzeug==2.0.3
Flask-Migrate==4.0.7
Flask-HTTPAuth==4.8.0